set-migration-files

+++ db/migrate/20230830103033_remove_requires_student_action_since_from_mentor_discussions.rb
@@ -1,7 +1,8 @@
 class RemoveRequiresStudentActionSinceFromMentorDiscussions < ActiveRecord::Migration[7.0]
   def change
     return if Rails.env.production? 
-    
-    remove_column :mentor_discussions, :requires_student_action_since
-  end
+    if column_exists?(:mentor_discussions, :requires_student_action_since)
+      remove_column :mentor_discussions, :requires_student_action_since
+    end  
+    end
 end
